fix!: Fix header value precision and HDU vector dispatch, add `Base.size(::HDU)`

+++ test/primary_hdu_tests.jl
@@ -140,11 +140,11 @@
                     "NAXIS2  =                    3                                                  "),
                    ("BZERO", 1.0f0, "",
                     "BZERO   =                  1.0                                                  "),
-                   ("BSCALE", 0.1f0, "",
+                   ("BSCALE", 0.1, "",
                     "BSCALE  =                  0.1                                                  ")])
 
     @test (ndims(hdu.data) == 2 && size(hdu.data) == (3, 3) && length(hdu.data) == 9 &&
-           eltype(hdu.data) == Float32 && all(hdu.data .== 1.1f0))
+           eltype(hdu.data) == Float64 && all(hdu.data .== 1.1))
 
     #  test Primary type with scale == false
     data = ones(Int32, (3,3))
@@ -173,7 +173,7 @@
                     "NAXIS2  =                    3                                                  "),
                    ("BZERO", 1.0f0, "",
                     "BZERO   =                  1.0                                                  "),
-                   ("BSCALE", 0.1f0, "",
+                   ("BSCALE", 0.1, "",
                     "BSCALE  =                  0.1                                                  ")])
 
     @test (ndims(hdu.data) == 2 && size(hdu.data) == (3, 3) && length(hdu.data) == 9 &&
@@ -210,11 +210,11 @@
                     "NAXIS2  =                    3                                                  "),
                    ("BZERO", 1.0f0, "",
                     "BZERO   =                  1.0                                                  "),
-                   ("BSCALE", 0.1f0, "",
+                   ("BSCALE", 0.1, "",
                     "BSCALE  =                  0.1                                                  ")])
 
     @test (ndims(hdu.data) == 2 && size(hdu.data) == (3, 3) && length(hdu.data) == 9 &&
-           eltype(hdu.data) == Float32 && all(hdu.data .== 1.1f0))
+           eltype(hdu.data) == Float64 && all(hdu.data .== 1.1))
 
     rm(temppath)
